refactor(nextjs): Consolidate game dialog and scoresheet creation workflow
- Remove separate add game store and provider
- Integrate scoresheet and game creation into a single dialog
- Simplify state management using React hooks
- Move schemas and form logic directly into components
- Remove unnecessary routing for scoresheet creation
- Enhance user experience with inline scoresheet configuration
